5. State machine based logic
1. Rebuild god class into state machine
2. Added Enhancement system (upgrades track, influence on turrets parameters)
3. Added Resource system, Reward generator - let player receive reward after level
4. Added Purchase system, lets player upgrade turrets for recourses
5. Advertisement system (implemented SDK class for level end multiplier)
6. Many UI improvements (Enhancement menu, level end menu)
7. Get rid of the damn bullet tracer problem (never scale down parent object when bullet flying)

TODO
1. NavMesh for enemy movement
2. New Terrains
3. Setup Menu
4. Sound
5. Leaderboard
6. Whatever...

+++ Assets/Scripts/Turret/Turrets/Turret.cs
@@ -16,7 +16,7 @@ public class Turret : MonoBehaviour
     public UnityAction EnemyFound;
 
     public IWeaponSystem WeaponSystem => _weaponSystem;
-    public int TurretLevel { get; private set; }
+    public float TurretLevel { get; private set; }
     
     private void Awake()
     {
@@ -25,7 +25,7 @@ public class Turret : MonoBehaviour
         EnemyFound += AttackEnemy;  
     }
 
-    public void Initialize(EnemyTrackSystem enemyTrackSystem, int turretLevel)
+    public void Initialize(EnemyTrackSystem enemyTrackSystem, float turretLevel)
     {
         _enemyTrackSystem = enemyTrackSystem;
         TurretLevel = turretLevel;
