+++ Assets/Scripts/Utils/Interfaces/IFactory.cs
@@ -1,4 +1,4 @@
 public interface IFactory<T>
 {
-    public T Build(int Level);
+    public T Build(float Level);
 }
